put helpers on model
