+++ packages/database/src/models/ImagingRequest/ImagingRequest.ts
@@ -1,4 +1,4 @@
-import { DataTypes, Sequelize } from 'sequelize';
+import { DataTypes, literal, Sequelize } from 'sequelize';
 import {
   IMAGING_REQUEST_STATUS_TYPES,
   IMAGING_TYPES_VALUES,
@@ -39,6 +39,8 @@ export class ImagingRequest extends Model {
   declare completedById?: string;
   declare locationGroupId?: string;
   declare locationId?: string;
+  /** Present when selected via {@link getApprovedAttribute}; not a DB column. */
+  declare approved?: boolean | null;
 
   declare areas: ImagingRequestArea[];
   declare encounter?: Encounter;
@@ -148,6 +150,73 @@ export class ImagingRequest extends Model {
 
   static getListReferenceAssociations() {
     return ['requestedBy', 'areas', 'results'];
+  }
+
+  /**
+   * Computed invoice-approval status for use in `attributes.include`.
+   * ImagingRequestArea items take precedence, then ImagingRequest items.
+   * @returns {[import('sequelize').Literal, 'approved']}
+   */
+  static getApprovedAttribute() {
+    return [
+      literal(`(
+        SELECT COALESCE(
+          -- ImagingRequestArea invoice items take precedence (NULL if none exist)
+          (
+            SELECT BOOL_AND(ii.approved)
+            FROM imaging_request_areas ira
+            INNER JOIN invoice_items ii ON ii.source_record_id = ira.id::text
+              AND ii.source_record_type = 'ImagingRequestArea'
+              AND ii.deleted_at IS NULL
+            WHERE ira.imaging_request_id = "ImagingRequest".id
+              AND ira.deleted_at IS NULL
+            HAVING COUNT(*) > 0
+          ),
+          -- ImagingRequest invoice items (used if area items returned NULL)
+          (
+            SELECT BOOL_AND(ii.approved)
+            FROM invoice_items ii
+            WHERE ii.source_record_id = "ImagingRequest".id::text
+              AND ii.source_record_type = 'ImagingRequest'
+              AND ii.deleted_at IS NULL
+            HAVING COUNT(*) > 0
+          )
+        )
+      )`),
+      'approved',
+    ];
+  }
+
+  /**
+   * Build a Sequelize order clause, with special handling for computed literal fields.
+   */
+  static getOrder(
+    orderBy?: string,
+    order: string = 'ASC',
+    options: { literalSortKeys?: string[] } = {},
+  ) {
+    if (!orderBy) return undefined;
+
+    const orderDirection = order.toUpperCase() === 'ASC' ? 'ASC' : 'DESC';
+    const { literalSortKeys = [] } = options;
+
+    let nullPosition;
+    if (orderBy === 'approved') {
+      nullPosition = 'NULLS LAST';
+    } else if (literalSortKeys.includes(orderBy)) {
+      nullPosition = orderDirection === 'ASC' ? 'NULLS FIRST' : 'NULLS LAST';
+    }
+
+    return [[...orderBy.split('.'), `${orderDirection}${nullPosition ? ` ${nullPosition}` : ''}`]];
+  }
+
+  forResponse() {
+    const values = super.forResponse();
+    // Base forResponse omits nulls; keep approved (including null) when it was selected
+    if ('approved' in this.dataValues) {
+      values.approved = this.get('approved') ?? null;
+    }
+    return values;
   }
 
   static initRelations(models: Models) {
